requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
requests==2.25.1
APScheduler==3.7.0
Werkzeug==1.0.1
pytz==2021.1
python-dotenv==0.15.0